made interns dynamic, put careers page components in separate folder, made pagination scroll up when clicked
