+++ ybt_website/src/components/careers_page/SearchBar.jsx
